[Fixes #1301] Auto Correct Grading Status
I tried pushing (through Git) the commit we worked on during the meeting. However, it gave me tslint errors due to variable shadowing of ```answer```. So I changed it to ```eachAnswer```, and it is still giving me errors. I figured I'd push the commit using Github so that you can take a look at it.