fix(certifications): Add image preview to certifications add/edit page
Add an image preview to the template to display the certificate
template. The preview is only shown if it's not the default
placeholder image. Also, change the 'Upload' button text to
'Change Template' if an image is already present. This provides
visual feedback to the user and resolves the issue of the
attachment appearing to be lost on page refresh.